chore: Clean up utils, helpers, PT serializer props
