+++ svelte-app/src/components/empty-content.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { t } from '$i18n';
+  import { t } from '$lib/i18n';
 
   import Icon from '$components/icon.svelte';
 </script>
