+++ svelte-app/src/components/nav/base-link.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { createEventDispatcher } from 'svelte';
 
-  import { linkTo, t } from '$i18n';
+  import { linkTo, t } from '$lib/i18n';
 
   import Hoverable from '$components/hoverable.svelte';
