+++ svelte-app/src/components/portable-text/footnote.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { t } from '$lib/helpers/i18n';
+  import { t } from '$lib/i18n';
 
   import Tooltip from '$components/tooltip.svelte';
