+++ svelte-app/src/components/portable-text/serializers/custom-paragraph.svelte
@@ -1,3 +1,11 @@
+<script lang="ts">
+  import type { BlockComponentProps } from '@portabletext/svelte';
+
+  export let portableText: BlockComponentProps;
+
+  $: ({ value: _value } = portableText);
+</script>
+
 <p><slot /></p>
 
 <style lang="scss">
