+++ svelte-app/src/components/portable-text/serializers/divider.svelte
@@ -1,5 +1,11 @@
 <script lang="ts">
   import Divider from '$components/divider.svelte';
+
+  import type { CustomBlockComponentProps } from '@portabletext/svelte';
+
+  export let portableText: CustomBlockComponentProps;
+
+  $: ({ value: _value } = portableText);
 </script>
 
 <Divider />
