+++ svelte-app/src/components/portable-text/serializers/list-item.svelte
@@ -1,1 +1,9 @@
+<script lang="ts">
+  import type { ListItemComponentProps } from '@portabletext/svelte';
+
+  export let portableText: ListItemComponentProps;
+
+  $: ({ value: _value } = portableText);
+</script>
+
 <li><slot /></li>
