+++ svelte-app/src/components/portable-text/serializers/null-mark.svelte
@@ -1,1 +1,9 @@
+<script lang="ts">
+  import type { MarkComponentProps } from '@portabletext/svelte';
+
+  export let portableText: MarkComponentProps;
+
+  $: ({ value: _value } = portableText);
+</script>
+
 <slot />
