+++ svelte-app/src/components/portable-text/serializers/ol-wrapper.svelte
@@ -1,3 +1,11 @@
+<script lang="ts">
+  import type { ListComponentProps } from '@portabletext/svelte';
+
+  export let portableText: ListComponentProps;
+
+  $: ({ value: _value } = portableText);
+</script>
+
 <ol class="pt-ol"><slot /></ol>
 
 <style lang="scss">
