+++ svelte-app/src/components/portable-text/serializers/ul-wrapper.svelte
@@ -1,3 +1,11 @@
+<script lang="ts">
+  import type { ListComponentProps } from '@portabletext/svelte';
+
+  export let portableText: ListComponentProps;
+
+  $: ({ value: _value } = portableText);
+</script>
+
 <ul class="pt-ul"><slot /></ul>
 
 <style lang="scss">
