+++ svelte-app/src/routes/[[lang=lang]]/work/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
-  import { pageTitle } from '$helpers/navigation';
-  import { t } from '$i18n';
+  import { t } from '$lib/i18n.js';
+  import { pageTitle } from '$lib/navigation';
 
   import Timeline from '$components/about/timeline.svelte';
   import Divider from '$components/divider.svelte';
