+++ svelte-app/src/routes/[[lang=lang]]/work/[slug]/+page.ts
@@ -1,7 +1,7 @@
 import { DEFAULT_APP_LANG } from '$lib/consts';
 import { fetchRepoStats, handleLoadError } from '$lib/data';
-import { buildImageUrl, getCrop } from '$lib/helpers/image';
 import Logger from '$lib/logger';
+import { buildImageUrl, getCrop } from '$lib/sanity';
 import { findOne } from '$lib/store';
 
 import type { PageLoad } from './$types';
