+++ svelte-app/tsconfig.json
@@ -3,7 +3,6 @@
   "compilerOptions": {
     "module": "es2022",
     "allowJs": true,
-    "baseUrl": ".",
     "checkJs": true,
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
@@ -11,25 +10,6 @@
     "noImplicitAny": true,
     "noImplicitThis": true,
     "noEmit": true,
-    "paths": {
-      "@colours": ["src/styles/colours"],
-      "@breakpoints": ["src/styles/breakpoints"],
-      "@utils": ["src/styles/utils"],
-      "@vars": ["src/styles/vars"],
-      "$lib": ["src/lib"],
-      "$lib/*": ["src/lib/*"],
-      "$components/*": ["src/components/*"],
-      "$helpers/*": ["src/lib/helpers/*"],
-      "$stores/*": ["src/stores/*"],
-      "$langs/*": ["src/languages/*"],
-      "$i18n": ["src/lib/helpers/i18n"],
-      "$routes/*": ["src/routes/*"],
-      "$assets/*": ["src/static/*"],
-      "$tests/*": ["tests/*"],
-      "$types": ["types/app/index.d.ts"],
-      "$types/*": ["types/app/*"],
-      "*": ["types/*"]
-    },
     "removeComments": true,
     "rootDir": "./",
     "resolveJsonModule": true,
